+++ Get-RecentBitly.ps1
@@ -1,0 +1,1 @@
+Get-BitlyGroupLinks -Size 5 | Select-object Created,ID,CustomLinks,Title,Tags,Url